v 1.3 changes
In OfferCredit:
worker continue to offer credit the next working day.

In OrderCredit:
set the deafualt language to german. If you want english, scroll down to
line 220 and remove -- in front of the english texts and write -- in
front of the german texts.

In idlelib.lua:
Changed the selection mode of sims to choose a bank.
Before it was "the nearest and most attractive bank will get all
customers".  Now it is a percent chance. The best bank has a higher
chance to get customers.